Created my submitted assignment page and how it works page

+++ src/Layouts/AssignmentDetails.jsx
@@ -4,8 +4,8 @@
  * ============================
  * [Done] 1. By clicking on the view assignment button the user will be navigated to the assignment details page (this page will be private) where the user will see the assignment details which are provided while the assignment was created. And the “Take assignment” button will be shown.
  * [Done] 2. Users are able to submit an assignment by clicking on the “Take assignment” button and it will open a modal [Or you may open a new page,it's up to you] with the assignment submission form.
- * 3. The assignment submission form will have input fields for PDF link submission and another text area for giving a quick note text.
- * 4. By default every submitted assignment will be in pending status . And save the user email with the submitted assignment so that it can be determined who has submitted it.
+ * [Done] 3. The assignment submission form will have input fields for PDF link submission and another text area for giving a quick note text.
+ * [Done] 4. By default every submitted assignment will be in pending status . And save the user email with the submitted assignment so that it can be determined who has submitted it.
  * 
 
  */
@@ -104,18 +104,31 @@ const AssignmentDetails = () => {
   }, []);
 
   const handleDeleteAssignment = () => {
-    Swal.fire({
-      title: "Do you want to delete this assignment?",
-      showCancelButton: true,
-      confirmButtonText: "Delete",
-      confirmButtonColor: "red",
-      focusConfirm: false,
-      focusCancel: true,
-    }).then((result) => {
-      if (result.isConfirmed) {
-        deleteAssignment();
-      }
-    });
+    if (user?.email === author?.email) {
+      Swal.fire({
+        icon: "question",
+        title: "Are you sure?",
+        text: "Do you want to delete this assignment?",
+        showCancelButton: true,
+        confirmButtonText: "Delete",
+        confirmButtonColor: "red",
+        focusConfirm: false,
+        focusCancel: true,
+      }).then((result) => {
+        if (result.isConfirmed) {
+          deleteAssignment();
+        }
+      });
+    } else {
+      // showToast("error", "You can only edit your won assignments.");
+      Swal.fire({
+        icon: "error",
+        title: "Forbidden!",
+        text: "You can only delete your won assignments.",
+        showConfirmButton: false,
+        showCloseButton: true,
+      });
+    }
   };
 
   const deleteAssignment = async () => {
